[7696] Fixed compile warnings Author: AlexDereka
--HG--
branch : trunk

+++ src/game/Object.h
@@ -203,7 +203,7 @@ class TRINITY_DLL_SPEC Object
             return *(((uint8*)&m_uint32Values[ index ])+offset);
         }
 
-        uint8 GetUInt16Value( uint16 index, uint8 offset) const
+        uint16 GetUInt16Value( uint16 index, uint8 offset) const
         {
             ASSERT( index < m_valuesCount || PrintIndexError( index , false) );
             ASSERT( offset < 2 );
